Fix compile with QT_NO_NARROWING_CONVERSIONS_IN_CONNECT flags
Summary: Fix compile with QT_NO_NARROWING_CONVERSIONS_IN_CONNECT flags

Reviewers: mgallien

Reviewed By: mgallien

Differential Revision: https://phabricator.kde.org/D15588

+++ src/audiowrapper.cpp
@@ -142,7 +142,7 @@ void AudioWrapper::stop()
     d->mPlayer.stop();
 }
 
-void AudioWrapper::seek(int position)
+void AudioWrapper::seek(qint64 position)
 {
     d->mPlayer.setPosition(position);
 }
